第七天:es数据导入搜索自动补全

+++ ly-item/ly-item-service/src/main/java/com/leyou/item/service/SkuService.java
@@ -13,5 +13,5 @@ public interface SkuService extends IService<Sku> {
 
     List<SkuDTO> querySkuById(List<Long> ids);
 
-    List<SkuDTO> querySkuBySpuId(Integer id);
+    List<SkuDTO> querySkuBySpuId(Long id);
 }
